fix: some changes to make asset state work
